+++ packages/api/api-generator/src/commands/generate/generateCrud/generate-crud.ts
@@ -229,6 +229,7 @@ function generatePaginatedDto({ generatorOptions, metadata }: { generatorOptions
 
 function generateArgsDto({ generatorOptions, metadata }: { generatorOptions: CrudGeneratorOptions; metadata: EntityMetadata<any> }): string {
     const { classNameSingular, fileNameSingular } = buildNameVariants(metadata);
+<<<<<<< HEAD
     const {
         scopeProp,
         argsClassName,
@@ -240,6 +241,10 @@ function generateArgsDto({ generatorOptions, metadata }: { generatorOptions: Cru
         crudSortProps,
         targetDirectory,
     } = buildOptions(metadata, generatorOptions);
+=======
+    const { scopeProp, argsClassName, hasSearchArg, hasSortArg, hasFilterArg, dedicatedResolverArgProps, hasPositionProp, crudSortProps, hasPaging } =
+        buildOptions(metadata, generatorOptions);
+>>>>>>> main
     const imports: Imports = [];
     if (scopeProp && scopeProp.targetMeta) {
         imports.push(generateEntityImport(scopeProp.targetMeta, `${targetDirectory}/dto`));
@@ -256,14 +261,14 @@ function generateArgsDto({ generatorOptions, metadata }: { generatorOptions: Cru
     const argsOut = `import { ArgsType, Field, IntersectionType, registerEnumType, ID } from "@nestjs/graphql";
     import { Type } from "class-transformer";
     import { IsOptional, IsString, ValidateNested, IsEnum, IsUUID } from "class-validator";
-    import { OffsetBasedPaginationArgs, SortDirection } from "@comet/cms-api";
-    import { ${classNameSingular}Filter } from "./${fileNameSingular}.filter";
-    import { ${classNameSingular}Sort, ${classNameSingular}SortField } from "./${fileNameSingular}.sort";
+    ${hasPaging ? `import { OffsetBasedPaginationArgs, SortDirection } from "@comet/cms-api";` : ``}
+    ${hasPaging && hasFilterArg ? `import { ${classNameSingular}Filter } from "./${fileNameSingular}.filter";` : ``}
+    ${hasPaging && hasSortArg ? `import { ${classNameSingular}Sort, ${classNameSingular}SortField } from "./${fileNameSingular}.sort";` : ``}
 
     ${generateImportsCode(imports)}
 
     @ArgsType()
-    export class ${argsClassName} extends OffsetBasedPaginationArgs {
+    export class ${argsClassName}${hasPaging ? ` extends OffsetBasedPaginationArgs` : ""} {
         ${
             scopeProp
                 ? `
@@ -291,7 +296,7 @@ function generateArgsDto({ generatorOptions, metadata }: { generatorOptions: Cru
             .join("")}
 
         ${
-            hasSearchArg
+            hasPaging && hasSearchArg
                 ? `
         @Field({ nullable: true })
         @IsOptional()
@@ -302,7 +307,7 @@ function generateArgsDto({ generatorOptions, metadata }: { generatorOptions: Cru
         }
 
         ${
-            hasFilterArg
+            hasPaging && hasFilterArg
                 ? `
         @Field(() => ${classNameSingular}Filter, { nullable: true })
         @ValidateNested()
@@ -314,7 +319,7 @@ function generateArgsDto({ generatorOptions, metadata }: { generatorOptions: Cru
         }
 
         ${
-            hasSortArg
+            hasPaging && hasSortArg
                 ? `
         @Field(() => [${classNameSingular}Sort], { ${defaultSortField === null ? "nullable: true" : `defaultValue: [{ field: ${classNameSingular}SortField.${defaultSortField}, direction: SortDirection.ASC }]`} })
         @ValidateNested({ each: true })
@@ -813,7 +818,12 @@ function generateResolver({ generatorOptions, metadata }: { generatorOptions: Cr
         positionGroupProps,
         hasDeletedAtProp,
         dedicatedResolverArgProps,
+<<<<<<< HEAD
         targetDirectory,
+=======
+        hasPaging,
+        hasArgsClass,
+>>>>>>> main
     } = buildOptions(metadata, generatorOptions);
 
     const imports: Imports = [];
@@ -890,8 +900,8 @@ function generateResolver({ generatorOptions, metadata }: { generatorOptions: Cr
 
     ${hasPositionProp ? `import { ${classNamePlural}Service } from "./${fileNamePlural}.service";` : ``}
     import { ${classNameSingular}Input, ${classNameSingular}UpdateInput } from "./dto/${fileNameSingular}.input";
-    import { Paginated${classNamePlural} } from "./dto/paginated-${fileNamePlural}";
-    import { ${argsClassName} } from "./dto/${argsFileName}";
+    ${hasPaging ? `import { Paginated${classNamePlural} } from "./dto/paginated-${fileNamePlural}";` : ""}
+    ${hasArgsClass ? `import { ${argsClassName} } from "./dto/${argsFileName}";` : ""}
     ${generateImportsCode(imports)}
 
     ${payloadObjectTypes.code}
@@ -939,13 +949,14 @@ function generateResolver({ generatorOptions, metadata }: { generatorOptions: Cr
         ${
             generatorOptions.list
                 ? `
-        @Query(() => Paginated${classNamePlural})
+        @Query(() => ${hasPaging ? `Paginated${classNamePlural}` : `[${metadata.className}]`})
         ${dedicatedResolverArgProps
             .map((dedicatedResolverArgProp) => {
                 return `@AffectedEntity(${dedicatedResolverArgProp.targetMeta?.className}, { idArg: "${dedicatedResolverArgProp.name}" })`;
             })
             .join("")}
         async ${instanceNameSingular != instanceNamePlural ? instanceNamePlural : `${instanceNamePlural}List`}(
+<<<<<<< HEAD
             @Args() {${Object.entries({
                 scope: !!scopeProp,
                 ...dedicatedResolverArgProps.reduce(
@@ -965,8 +976,34 @@ function generateResolver({ generatorOptions, metadata }: { generatorOptions: Cr
                 .map(([key]) => key)
                 .join(", ")}}: ${argsClassName}
         ): Promise<Paginated${classNamePlural}> {
+=======
+            ${
+                hasArgsClass
+                    ? `@Args() {${Object.entries({
+                          scope: !!scopeProp,
+                          ...dedicatedResolverArgProps.reduce(
+                              (acc, dedicatedResolverArgProp) => {
+                                  acc[dedicatedResolverArgProp.name] = true;
+                                  return acc;
+                              },
+                              {} as Record<string, boolean>,
+                          ),
+                          search: hasPaging && !!hasSearchArg,
+                          filter: hasPaging && !!hasFilterArg,
+                          sort: hasPaging && !!hasSortArg,
+                          offset: hasPaging,
+                          limit: hasPaging,
+                      })
+                          .filter(([key, use]) => use)
+                          .map(([key]) => key)
+                          .join(", ")}}: ${argsClassName}`
+                    : ""
+            }
+            ${hasOutputRelations ? `${hasArgsClass ? `,` : ""} @Info() info: GraphQLResolveInfo` : ""}
+        ): Promise<${hasPaging ? `Paginated${classNamePlural}` : `${metadata.className}[]`}> {
+>>>>>>> main
             const where${
-                hasSearchArg || hasFilterArg
+                hasPaging && (hasSearchArg || hasFilterArg)
                     ? ` = gqlArgsToMikroOrmQuery({ ${hasSearchArg ? `search, ` : ""}${hasFilterArg ? `filter, ` : ""} }, this.entityManager.getMetadata(${metadata.className}));`
                     : `: ObjectQuery<${metadata.className}> = {}`
             }
@@ -977,18 +1014,43 @@ function generateResolver({ generatorOptions, metadata }: { generatorOptions: Cr
                 })
                 .join("\n")}
 
+<<<<<<< HEAD
             const options: FindOptions<${metadata.className}> = { offset, limit };
+=======
+            ${
+                hasOutputRelations
+                    ? `const fields = extractGraphqlFields(info${hasPaging ? `, { root: "nodes" }` : ""});
+            const populate: string[] = [];`
+                    : ""
+            }
+            ${[...outputRelationManyToOneProps, ...outputRelationOneToManyProps, ...outputRelationManyToManyProps, ...outputRelationOneToOneProps]
+                .map(
+                    (r) =>
+                        `if (fields.includes("${r.name}")) {
+                            populate.push("${r.name}");
+                        }`,
+                )
+                .join("\n")}
+
+            ${hasOutputRelations ? `// eslint-disable-next-line @typescript-eslint/no-explicit-any` : ""}
+            const options: FindOptions<${metadata.className}${hasOutputRelations ? `, any` : ""}> = { ${hasPaging ? `offset, limit${hasOutputRelations ? `, populate` : ""}` : `${hasOutputRelations ? `populate` : ""}`}};
+>>>>>>> main
 
             ${
-                hasSortArg
+                hasPaging && hasSortArg
                     ? `if (sort) {
                 options.orderBy = gqlSortToMikroOrmOrderBy(sort);
             }`
                     : ""
             }
 
-            const [entities, totalCount] = await this.entityManager.findAndCount(${metadata.className}, where, options);
-            return new Paginated${classNamePlural}(entities, totalCount);
+            ${
+                hasPaging
+                    ? `const [entities, totalCount] = await this.entityManager.findAndCount(${metadata.className}, where, options);
+            return new Paginated${classNamePlural}(entities, totalCount);`
+                    : `const entities = await this.entityManager.find(${metadata.className}, where, options);
+            return entities;`
+            }
         }
         `
                 : ""
@@ -1198,29 +1260,42 @@ export async function generateCrud(generatorOptionsParam: CrudGeneratorOptions, 
     const generatedFiles: GeneratedFile[] = [];
 
     const { fileNameSingular, fileNamePlural } = buildNameVariants(metadata);
-    const { hasFilterArg, hasSortArg, argsFileName, hasPositionProp } = buildOptions(metadata, generatorOptions);
+    const { hasFilterArg, hasSortArg, argsFileName, hasPositionProp, hasPaging, hasArgsClass } = buildOptions(metadata, generatorOptions);
 
     async function generateCrudResolver(): Promise<GeneratedFile[]> {
+<<<<<<< HEAD
         if (hasFilterArg) {
             generatedFiles.push(...generateFilterDto({ generatorOptions, metadata }));
+=======
+        if (hasPaging && hasFilterArg) {
+            generatedFiles.push({
+                name: `dto/${fileNameSingular}.filter.ts`,
+                content: generateFilterDto({ generatorOptions, metadata }),
+                type: "filter",
+            });
+>>>>>>> main
         }
-        if (hasSortArg) {
+        if (hasPaging && hasSortArg) {
             generatedFiles.push({
                 name: `dto/${fileNameSingular}.sort.ts`,
                 content: generateSortDto({ generatorOptions, metadata }),
                 type: "sort",
             });
         }
-        generatedFiles.push({
-            name: `dto/paginated-${fileNamePlural}.ts`,
-            content: generatePaginatedDto({ generatorOptions, metadata }),
-            type: "sort",
-        });
-        generatedFiles.push({
-            name: `dto/${argsFileName}.ts`,
-            content: generateArgsDto({ generatorOptions, metadata }),
-            type: "args",
-        });
+        if (hasPaging) {
+            generatedFiles.push({
+                name: `dto/paginated-${fileNamePlural}.ts`,
+                content: generatePaginatedDto({ generatorOptions, metadata }),
+                type: "sort",
+            });
+        }
+        if (hasArgsClass) {
+            generatedFiles.push({
+                name: `dto/${argsFileName}.ts`,
+                content: generateArgsDto({ generatorOptions, metadata }),
+                type: "args",
+            });
+        }
         if (hasPositionProp) {
             generatedFiles.push({
                 name: `${fileNamePlural}.service.ts`,