+++ packages/site/site-react/src/blocks/PhoneLinkBlock.tsx
@@ -1,35 +1,45 @@
+<<<<<<< HEAD
 import { type AnchorHTMLAttributes, cloneElement, type DetailedHTMLProps, type ReactElement } from "react";
+=======
+import { type AnchorHTMLAttributes, cloneElement, type ReactElement } from "react";
+>>>>>>> main
 
 import { type PhoneLinkBlockData } from "../blocks.generated";
 import { type PropsWithData } from "./PropsWithData";
 
-interface PhoneLinkBlockProps extends PropsWithData<PhoneLinkBlockData> {
+interface PhoneLinkBlockProps extends PropsWithData<PhoneLinkBlockData>, Omit<AnchorHTMLAttributes<HTMLAnchorElement>, "href"> {
     children: ReactElement;
-    title?: string;
-    className?: string;
     legacyBehavior?: boolean;
 }
 
+<<<<<<< HEAD
 export const PhoneLinkBlock = ({ data: { phone }, children, title, className, legacyBehavior }: PhoneLinkBlockProps): React.ReactNode => {
+=======
+export const PhoneLinkBlock = ({ data: { phone }, children, legacyBehavior, ...anchorProps }: PhoneLinkBlockProps) => {
+>>>>>>> main
     if (!phone) {
         if (legacyBehavior) {
             return children;
         }
 
-        return <span className={className}>{children}</span>;
+        return <span className={anchorProps.className}>{children}</span>;
     }
 
     const href = `tel:${phone}`;
 
     if (legacyBehavior) {
+<<<<<<< HEAD
         return cloneElement(
             children as ReactElement<Pick<DetailedHTMLProps<AnchorHTMLAttributes<HTMLAnchorElement>, HTMLAnchorElement>, "href" | "title">>,
             { href, title },
         );
+=======
+        return cloneElement(children, { ...anchorProps, href });
+>>>>>>> main
     }
 
     return (
-        <a href={href} title={title} className={className}>
+        <a {...anchorProps} href={href}>
             {children}
         </a>
     );