+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @comet/eslint-config
 
+## 8.17.1
+
+### Patch Changes
+
+- @comet/eslint-plugin@8.17.1
+
+## 8.17.0
+
+### Patch Changes
+
+- @comet/eslint-plugin@8.17.0
+
+## 8.16.0
+
+### Patch Changes
+
+- @comet/eslint-plugin@8.16.0
+
 ## 8.15.0
 
 ### Patch Changes